Fix AI token not refreshing when config is updated
Additionally fixed false positive for rule 2 by implementing levenshtein distance calculation, because the AI was often suggesting a simple reordering of words or single-character difference (e.g. AI didn't like the period at the end of the original message, so the suggestion simply removed the period) with regards to the original. Now the GradedCommitDisplay removes the rule violations for Rule 2 if the difference in the suggestion and the original is only 1 character.
Rule violations are not removed in case of the body strings being similar, only for header strings similarities.